+++ service/src/main/java/eu/europa/ec/fisheries/uvms/movement/service/mapper/SegmentMapper.java
@@ -9,14 +9,14 @@ public class SegmentMapper {
     public static SegmentDTO mapToSegmentDTO(Segment segment){
         SegmentDTO dto = new SegmentDTO();
         dto.setId(segment.getId());
-        dto.setCourseOverGround(segment.getCourseOverGround());
+        dto.setCourseOverGround((double)segment.getCourseOverGround());
         dto.setDistance(segment.getDistance());
         dto.setDuration(segment.getDuration());
         dto.setFromMovement(segment.getFromMovement().getId().toString());
         dto.setToMovement(segment.getToMovement().getId().toString());
         dto.setLocation(WKTUtil.getWktLineStringFromSegment(segment));
         dto.setSegmentCategory(segment.getSegmentCategory());
-        dto.setSpeedOverGround(segment.getSpeedOverGround());
+        dto.setSpeedOverGround((double)segment.getSpeedOverGround());
         dto.setTrack(segment.getTrack().getId());
         dto.setUpdated(segment.getUpdated());
         dto.setUpdatedBy(segment.getUpdatedBy());
