+++ service/src/main/java/eu/europa/ec/fisheries/uvms/movement/service/message/IncomingMovementMapper.java
@@ -18,8 +18,8 @@ public abstract class IncomingMovementMapper {
     public static Movement mapNewMovementEntity(IncomingMovement ic, String username) {
         Movement entity = new Movement();
 
-        entity.setSpeed(ic.getReportedSpeed());
-        entity.setHeading(ic.getReportedCourse());
+        entity.setSpeed(ic.getReportedSpeed().floatValue());
+        entity.setHeading(ic.getReportedCourse().floatValue());
         entity.setInternalReferenceNumber(ic.getInternalReferenceNumber());
         entity.setTripNumber(ic.getTripNumber());
         entity.setStatus(ic.getStatus());
@@ -94,12 +94,12 @@ public abstract class IncomingMovementMapper {
         md.setLatitude(movement.getLocation().getY());
         md.setMovementType(movement.getMovementType().value());
         if(movement.getFromSegment() != null) {
-            md.setCalculatedCourse(movement.getFromSegment().getCourseOverGround());
-            md.setCalculatedSpeed(movement.getFromSegment().getSpeedOverGround());
+            md.setCalculatedCourse((double)movement.getFromSegment().getCourseOverGround());
+            md.setCalculatedSpeed((double)movement.getFromSegment().getSpeedOverGround());
             md.setSegmentType(movement.getFromSegment().getSegmentCategory().value());
         }
-        md.setReportedCourse(movement.getHeading());
-        md.setReportedSpeed(movement.getSpeed());
+        md.setReportedCourse((double)movement.getHeading());
+        md.setReportedSpeed((double)movement.getSpeed());
         md.setPositionTime(movement.getTimestamp());
         md.setStatusCode(movement.getStatus());
         md.setTripNumber(movement.getTripNumber());
